+++ nodeconductor/core/views.py
@@ -198,54 +198,6 @@ def exception_handler(exc, context):
     return rf_exception_handler(exc, context)
 
 
-# XXX: Deprecated. Use SummaryQueryset + regular view instead.
-#      Check ResourceSymmaryVeiw for example.
-class BaseSummaryView(GenericViewSet):
-    params = []
-
-    def list(self, request):
-        qs = self.get_queryset(request)
-        qs = self.order_queryset(request, qs)
-        page = self.paginate_queryset(qs)
-        if page is not None:
-            return self.get_paginated_response(page)
-        return Response(qs)
-
-    def get_queryset(self, request):
-        def fetch_data(view_name, params):
-            response = request_api(request, view_name, params=params)
-            if not response.ok:
-                raise APIException(response.text)
-            return response
-
-        data = []
-        for url in self.get_urls(request):
-            params = self.get_params(request)
-            response = fetch_data(url, params)
-
-            json = response.json()
-            if response.total and response.total > len(json):
-                params['page_size'] = response.total
-            data += json
-        return data
-
-    def get_params(self, request):
-        params = {}
-        for key in self.params:
-            if key in request.query_params:
-                params[key] = request.query_params.get(key)
-        return params
-
-    def get_urls(self, request):
-        return []
-
-    def order_queryset(self, request, qs):
-        field = request.query_params.get('o')
-        if not field:
-            return qs
-        return sorted(qs, key=lambda x: x.get(field))
-
-
 class StateExecutorViewSet(mixins.StateMixin,
                            mixins.CreateExecutorMixin,
                            mixins.UpdateExecutorMixin,